First attempt at implementing MvcEnv.
This implementation has a bunch of issues, but I think is generally a
step in the right direction.

How to handle AddV type actions is still really unclear, and basically
wrong in this version.

This version also doesn't update the *_seq member variables of IEnv,
which is probably wrong too.